requirements.txt
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.4.0
numpy==1.20.2
panda==0.3.1
pandas==1.2.3
Pillow==8.1.2
pyirsdk==1.2.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.24.0
six==1.15.0
urllib3==1.25.11
